requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
pandas>=1.3.3
h5py>=3.4.0
matplotlib>=3.4.3
numpy>=1.21.2
Pillow>=8.3.2
opencv-python>=4.5.3